remove AbstractClient
