+++ airbase/download_api/__init__.py
@@ -1,4 +1,3 @@
-from .abstract_api_client import AbstractClient
 from .api_client import (
     Dataset,
     ParquetData,
